refactor(chip/icon/navbar): optimize CSS
