+++ src/components/chip/themes/shared/chip.bootstrap.scss
@@ -4,25 +4,8 @@
 $theme: $bootstrap;
 
 :host button {
-    border-radius: var-get($theme, 'border-radius');
-    background: var-get($theme, 'background');
-    color: var-get($theme, 'text-color');
-    height: var-get($theme, 'size');
-
-    &:hover {
-        background: var-get($theme, 'hover-background');
-        color: var-get($theme, 'hover-text-color');
-    }
-
     &:focus {
         outline: rem(2px) solid var-get($theme, 'focus-outline-color');
-        background: var-get($theme, 'focus-background');
-        color: var-get($theme, 'focus-text-color');
-    }
-
-    &[disabled] {
-        background: var-get($theme, 'disabled-background');
-        color: var-get($theme, 'disabled-text-color');
     }
 }
