+++ src/components/chip/themes/shared/chip.common.scss
