+++ src/components/chip/themes/shared/chip.fluent.scss
@@ -5,27 +5,6 @@ $theme: $fluent;
 
 :host button {
     @include type-style('subtitle-2');
-
-    border-radius: var-get($theme, 'border-radius');
-    background: var-get($theme, 'background');
-    color: var-get($theme, 'text-color');
-    outline: none;
-    height: var-get($theme, 'size');
-
-    &:hover {
-        background: var-get($theme, 'hover-background');
-        color: var-get($theme, 'hover-text-color');
-    }
-
-    &:focus {
-        background: var-get($theme, 'focus-background');
-        color: var-get($theme, 'focus-text-color');
-    }
-
-    &[disabled] {
-        background: var-get($theme, 'disabled-background');
-        color: var-get($theme, 'disabled-text-color');
-    }
 }
 
 :host([selected]) button:not([disabled]) {
