+++ src/components/icon/themes/shared/icon.common.scss
