--- src/components/icon/themes/shared/icon.indigo.scss
@@ -1,17 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    color: var-get($theme, 'color');
-}
-
-svg {
-    fill: currentcolor;
-}
-
-igc-icon,
-svg {
-    font-size: var-get($theme, 'size');
-}
