+++ src/components/navbar/themes/shared/navbar.bootstrap.scss
@@ -1,21 +1,8 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $bootstrap;
 
 :host {
-    background: var-get($theme, 'background');
-    color: var-get($theme, 'text-color');
     box-shadow: var(--ig-elevation-0);
     padding: 0 rem(8px);
-}
-
-::slotted(igc-icon) {
-    color: var-get($theme, 'idle-icon-color');
-}
-
-::slotted(igc-icon:hover) {
-    color: var-get($theme, 'hover-icon-color');
 }
 
 [part='base'] {
