+++ src/components/navbar/themes/shared/navbar.fluent.scss
@@ -1,21 +1,4 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $fluent;
-
-:host {
-    background: var-get($theme, 'background');
-    color: var-get($theme, 'text-color');
-    box-shadow: var(--ig-elevation-4);
-}
-
-::slotted(igc-icon) {
-    color: var-get($theme, 'idle-icon-color');
-}
-
-::slotted(igc-icon:hover) {
-    color: var-get($theme, 'hover-icon-color');
-}
 
 ::slotted(*) {
     @include type-style('subtitle-2') {
