+++ src/components/navbar/themes/shared/navbar.indigo.scss
@@ -1,18 +1,5 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $indigo;
 
 :host {
-    background: var-get($theme, 'background');
-    color: var-get($theme, 'text-color');
     box-shadow: var(--ig-elevation-0);
-}
-
-::slotted(igc-icon) {
-    color: var-get($theme, 'idle-icon-color');
-}
-
-::slotted(igc-icon:hover) {
-    color: var-get($theme, 'hover-icon-color');
 }
